Fix llvm Clang 10.0.1 compiler errors
#jira UE-100249
#rb Michael.Sartian

[CL 14391673 by Brandon Schaefer in 4.26 branch]

+++ Engine/Source/Editor/DetailCustomizations/Private/MaterialProxySettingsCustomizations.cpp
@@ -132,7 +132,7 @@ void FMaterialProxySettingsCustomizations::AddTextureSizeClamping(TSharedPtr<IPr
 	//     with 16 bytes per pixel.
 	//     See implimentations of ImageUtils.cpp :: GetRawData()  and ImageCore.cpp :: CopyTo()
 	//     11585 = Floor( Sqrt ( Max_int32 / 16)  )
-	const int32 TmpMaxSize = FMath::FloorToInt(FMath::Sqrt(MAX_int32 / 16));
+	const int32 TmpMaxSize = FMath::FloorToInt(FMath::Sqrt(static_cast<float>(MAX_int32) / 16));
 	const int32 MaxProxyTextureResolution = FMath::Min(TmpMaxSize, (int32)GetMax2DTextureDimension());
 	
 	const FString MaxTextureResolutionString = FString::FromInt(MaxProxyTextureResolution);
